services/yfinance-poll/requirements.txt
kafka-python
yfinance==0.2.24
boto3
protobuf
google